Create proper example code (Not yet used in docs. Will be part of future PRs)
